Added operations
Now you can delete tasks.

Refactoring:
made a function that will take input and convert it to uint8
instead of typing it each time.

Note:
This commit will be the last commit. I am getting bored of this
project.
I will move to a new one and maybe will come back to this one
later.